Gallery and firebase storage

+++ lib/firebase_firestore/Firestore.dart
@@ -1,9 +1,10 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_storage/firebase_storage.dart';
 class FirestoreOperation {
-  var fire = FirebaseFirestore.instance.collection('data');
+  var dbref = FirebaseFirestore.instance.collection('data');
 
-  setDataInFirestore({required String name, required String description, required String email, required var image,}) {
-    fire
+  addDataFirestore({required String name, required String description, required String email, required var image,}) {
+    dbref
         .add({
           'name': name,
           'description': description,
@@ -14,8 +15,8 @@ class FirestoreOperation {
         .catchError((error) => print("Failed to update user: $error"));
   }
 
-  updateFirestore({required String name, required String description, required String email, required String image, required String id}) {
-    fire.doc(id)
+  updateDataFirestore({required String name, required String description, required String email, required String image, required String id}) {
+    dbref.doc(id)
         .update({
           'name': name,
           'description': description,
@@ -26,12 +27,24 @@ class FirestoreOperation {
         .catchError((error) => print("Failed to update user: $error"));
   }
 
-  deleteDataFromFirestore({required String id}) {
-    fire
-        .doc(id)
-        .delete()
-        .then((value) => print("deleting successfully"))
-        .catchError((e) => print("$e \nerror while deleting the element"));
+  deleteDataFirestore({required String id,required var model,required bool fromUpdate}) async {
+
+    String filePath = model["image"].replaceAll(new RegExp(r'(\?alt).*'), '');
+
+    List split=filePath.split("data%2F");
+
+    FirebaseStorage.instance.ref("data").child(split[1]).delete()
+        .then((_) => print('Successfully deleted $filePath storage item' ))
+        .catchError((_){print("image not exist in the firebase storage");});
+
+    if(!fromUpdate){
+      dbref
+          .doc(id)
+          .delete()
+          .then((value) => print("deleting successfully"))
+          .catchError((e) => print("$e \nerror while deleting the element"));
+      print(id);
+    }
   }
 
   }
